rag/requirements.txt
python-dotenv
langchain
langchain-openai
langchain-community
langchain-pinecone
langchainhub
black